Remove commented out lines
