+++ src/pages/Home/Home.page.jsx
@@ -1,34 +1,11 @@
 import React, { useState } from 'react';
-// import { Link, useHistory } from 'react-router-dom';
-
-// import { useAuth } from '../../providers/Auth';
 import { VideoList, HomeTitle } from './Home.styles.jsx';
 import mock from './mock/mock.json'
 import VideoCard from '../../components/VideoCard'
+import { getReadableDate } from '../../utils/fns';
 
 function HomePage() {
   const [videoList, setVideoList] = useState(mock.items)
-  /* const history = useHistory();
-  const sectionRef = useRef(null);
-  const { authenticated, logout } = useAuth();
-
-  function deAuthenticate(event) {
-    event.preventDefault();
-    logout();
-    history.push('/');
-  } */
-
-  const getReadableDate = (timeStamp) => {
-    const date = new Date(timeStamp)
-    const options = {
-      weekday: 'long',
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric'
-    };
-
-    return date.toLocaleDateString(undefined, options)
-  }
 
   const getUsefullData = (videoData, idx) => {
     const { publishedAt, title, channelTitle,
@@ -54,23 +31,6 @@ function HomePage() {
       )}
     </VideoList>
     </section>
-    /* <section className="homepage" ref={sectionRef}>
-      <h1>Hello stranger!</h1>
-      {authenticated ? (
-        <>
-          <h2>Good to have you back</h2>
-          <span>
-            <Link to="/" onClick={deAuthenticate}>
-              ← logout
-            </Link>
-            <span className="separator" />
-            <Link to="/secret">show me something cool →</Link>
-          </span>
-        </>
-      ) : (
-        <Link to="/login">let me in →</Link>
-      )}
-    </section> */
   );
 }
